One MCMC per model with multiple outputs

Targets to run a JAGS model once with MCMC and save multiple outputs.

Usage

tar_jags(
  name,
  jags_files,
  parameters.to.save,
  data = list(),
  summaries = list(),
  summary_args = list(),
  n.cluster = 1,
  n.chains = 3,
  n.iter = 2000,
  n.burnin = as.integer(n.iter/2),
  n.thin = 1,
  jags.module = c("glm", "dic"),
  inits = NULL,
  RNGname = c("Wichmann-Hill", "Marsaglia-Multicarry", "Super-Duper", "Mersenne-Twister"),
  jags.seed = 1,
  stdout = NULL,
  stderr = NULL,
  progress.bar = "text",
  refresh = 0,
  draws = TRUE,
  summary = TRUE,
  dic = TRUE,
  tidy_eval = targets::tar_option_get("tidy_eval"),
  packages = targets::tar_option_get("packages"),
  library = targets::tar_option_get("library"),
  format = "qs",
  format_df = "fst_tbl",
  repository = targets::tar_option_get("repository"),
  error = targets::tar_option_get("error"),
  memory = targets::tar_option_get("memory"),
  garbage_collection = targets::tar_option_get("garbage_collection"),
  deployment = targets::tar_option_get("deployment"),
  priority = targets::tar_option_get("priority"),
  resources = targets::tar_option_get("resources"),
  storage = targets::tar_option_get("storage"),
  retrieval = targets::tar_option_get("retrieval"),
  cue = targets::tar_option_get("cue"),
  description = targets::tar_option_get("description")
)

Arguments

name

Symbol, base name for the collection of targets. Serves as a prefix for target names.

jags_files

Character vector of JAGS model files. If you supply multiple files, each model will run on the one shared dataset generated by the code in data. If you supply an unnamed vector, tools::file_path_sans_ext(basename(jags_files)) will be used as target name suffixes. If jags_files is a named vector, the suffixed will come from names(jags_files).

parameters.to.save

Model parameters to save, passed to R2jags::jags() or R2jags::jags.parallel(). See the argument documentation of the R2jags::jags() and R2jags::jags.parallel() help files for details.

data

Code to generate the data list for the JAGS model. Optionally include a .join_data element to join parts of the data to correspondingly named parameters in the summary output. See the vignettes for details.

summaries

List of summary functions passed to ... in posterior::summarize_draws() through $summary() on the CmdStanFit object.

summary_args

List of summary function arguments passed to .args in posterior::summarize_draws() through $summary() on the CmdStanFit object.

n.cluster

Number of parallel processes, passed to R2jags::jags() or R2jags::jags.parallel(). See the argument documentation of the R2jags::jags() and R2jags::jags.parallel() help files for details.

n.chains

Number of MCMC chains, passed to R2jags::jags() or R2jags::jags.parallel(). See the argument documentation of the R2jags::jags() and R2jags::jags.parallel() help files for details.

n.iter

Number if iterations (including warmup), passed to R2jags::jags() or R2jags::jags.parallel(). See the argument documentation of the R2jags::jags() and R2jags::jags.parallel() help files for details.

n.burnin

Number of warmup iterations, passed to R2jags::jags() or R2jags::jags.parallel(). See the argument documentation of the R2jags::jags() and R2jags::jags.parallel() help files for details.

n.thin

Thinning interval, passed to R2jags::jags() or R2jags::jags.parallel(). See the argument documentation of the R2jags::jags() and R2jags::jags.parallel() help files for details.

jags.module

Character vector of JAGS modules to load, passed to R2jags::jags() or R2jags::jags.parallel(). See the argument documentation of the R2jags::jags() and R2jags::jags.parallel() help files for details.

inits

Initial values of model parameters, passed to R2jags::jags() or R2jags::jags.parallel(). See the argument documentation of the R2jags::jags() and R2jags::jags.parallel() help files for details.

RNGname

Choice of random number generator, passed to R2jags::jags() or R2jags::jags.parallel(). See the argument documentation of the R2jags::jags() and R2jags::jags.parallel() help files for details.

jags.seed

Seeds to apply to JAGS, passed to R2jags::jags() or R2jags::jags.parallel(). See the argument documentation of the R2jags::jags() and R2jags::jags.parallel() help files for details.

stdout

Character of length 1, file path to write the stdout stream of the model when it runs. Set to NULL to print to the console. Set to R.utils::nullfile() to suppress stdout. Does not apply to messages, warnings, or errors.

stderr

Character of length 1, file path to write the stderr stream of the model when it runs. Set to NULL to print to the console. Set to R.utils::nullfile() to suppress stderr. Does not apply to messages, warnings, or errors.

progress.bar

Type of progress bar, passed to R2jags::jags() or R2jags::jags.parallel(). See the argument documentation of the R2jags::jags() and R2jags::jags.parallel() help files for details.

refresh

Frequency for refreshing the progress bar, passed to R2jags::jags() or R2jags::jags.parallel(). See the argument documentation of the R2jags::jags() and R2jags::jags.parallel() help files for details.

draws

Logical, whether to create a target for posterior draws. Saves draws as a compressed posterior::as_draws_df() tibble. Convenient, but duplicates storage.

summary

Logical, whether to create a target to store a small data frame of posterior summary statistics and convergence diagnostics.

dic

Logical, whether to create a target with deviance information criterion (DIC) results.

tidy_eval

Logical, whether to enable tidy evaluation when interpreting command and pattern. If TRUE, you can use the "bang-bang" operator !! to programmatically insert the values of global objects.

packages

Character vector of packages to load right before the target runs or the output data is reloaded for downstream targets. Use tar_option_set() to set packages globally for all subsequent targets you define.

library

Character vector of library paths to try when loading packages.

format

Character of length 1, storage format of the non-data-frame targets such as the JAGS data and any JAGS fit objects. Please choose an all=purpose format such as "qs" or "aws_qs" rather than a file format like "file" or a data frame format like "parquet". For more on storage formats, see the help file of targets::tar_target().

format_df

Character of length 1, storage format of the data frame targets such as posterior draws. We recommend efficient data frame formats such as "feather" or "aws_parquet". For more on storage formats, see the help file of targets::tar_target().

repository

Character of length 1, remote repository for target storage. Choices:

"local": file system of the local machine.

"aws": Amazon Web Services (AWS) S3 bucket. Can be configured with a non-AWS S3 bucket using the endpoint argument of tar_resources_aws(), but versioning capabilities may be lost in doing so. See the cloud storage section of https://books.ropensci.org/targets/data.html for details for instructions.

"gcp": Google Cloud Platform storage bucket. See the cloud storage section of https://books.ropensci.org/targets/data.html for details for instructions.

Note: if repository is not "local" and format is "file" then the target should create a single output file. That output file is uploaded to the cloud and tracked for changes where it exists in the cloud. The local file is deleted after the target runs.

error

Character of length 1, what to do if the target stops and throws an error. Options:

"stop": the whole pipeline stops and throws an error.

"continue": the whole pipeline keeps going.

"abridge": any currently running targets keep running, but no new targets launch after that. (Visit https://books.ropensci.org/targets/debugging.html to learn how to debug targets using saved workspaces.)

"null": The errored target continues and returns NULL. The data hash is deliberately wrong so the target is not up to date for the next run of the pipeline.

memory

Character of length 1, memory strategy. If "persistent", the target stays in memory until the end of the pipeline (unless storage is "worker", in which case targets unloads the value from memory right after storing it in order to avoid sending copious data over a network). If "transient", the target gets unloaded after every new target completes. Either way, the target gets automatically loaded into memory whenever another target needs the value. For cloud-based dynamic files (e.g. format = "file" with repository = "aws"), this memory strategy applies to the temporary local copy of the file: "persistent" means it remains until the end of the pipeline and is then deleted, and "transient" means it gets deleted as soon as possible. The former conserves bandwidth, and the latter conserves local storage.

garbage_collection

Logical, whether to run base::gc() just before the target runs.

deployment

Character of length 1. If deployment is "main", then the target will run on the central controlling R process. Otherwise, if deployment is "worker" and you set up the pipeline with distributed/parallel computing, then the target runs on a parallel worker. For more on distributed/parallel computing in targets, please visit https://books.ropensci.org/targets/crew.html.

priority

Numeric of length 1 between 0 and 1. Controls which targets get deployed first when multiple competing targets are ready simultaneously. Targets with priorities closer to 1 get dispatched earlier (and polled earlier in tar_make_future()).

resources

Object returned by tar_resources() with optional settings for high-performance computing functionality, alternative data storage formats, and other optional capabilities of targets. See tar_resources() for details.

storage

Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). Must be one of the following values:

"main": the target's return value is sent back to the host machine and saved/uploaded locally.

"worker": the worker saves/uploads the value.

"none": almost never recommended. It is only for niche situations, e.g. the data needs to be loaded explicitly from another language. If you do use it, then the return value of the target is totally ignored when the target ends, but each downstream target still attempts to load the data file (except when retrieval = "none").

If you select storage = "none", then the return value of the target's command is ignored, and the data is not saved automatically. As with dynamic files (format = "file") it is the responsibility of the user to write to the data store from inside the target.

The distinguishing feature of storage = "none" (as opposed to format = "file") is that in the general case, downstream targets will automatically try to load the data from the data store as a dependency. As a corollary, storage = "none" is completely unnecessary if format is "file".

retrieval

Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). Must be one of the following values:

"main": the target's dependencies are loaded on the host machine and sent to the worker before the target runs.

"worker": the worker loads the targets dependencies.

"none": the dependencies are not loaded at all. This choice is almost never recommended. It is only for niche situations, e.g. the data needs to be loaded explicitly from another language.

cue

An optional object from tar_cue() to customize the rules that decide whether the target is up to date.

description

Character of length 1, a custom free-form human-readable text description of the target. Descriptions appear as target labels in functions like tar_manifest() and tar_visnetwork(), and they let you select subsets of targets for the names argument of functions like tar_make(). For example, tar_manifest(names = tar_described_as(starts_with("survival model"))) lists all the targets whose descriptions start with the character string "survival model".

Value

tar_jags() returns list of target objects. See the "Target objects" section for background. The target names use the name argument as a prefix, and the individual elements of jags_files appear in the suffixes where applicable. As an example, the specific target objects returned by tar_jags(name = x, jags_files = "y.jags", ...) returns a list of targets::tar_target() objects:

x_file_y: reproducibly track the JAGS model file. Returns a character vector of length 1 with the path to the JAGS model file.

x_lines_y: read the contents of the JAGS model file for safe transport to parallel workers. Returns a character vector of lines in the model file.

x_data: run the R expression in the data argument to produce a JAGS dataset for the model. Returns a JAGS data list.

x_mcmc_y: run MCMC on the model and dataset. Returns an rjags object from R2jags with all the MCMC results.

x_draws_y: extract posterior samples from x_mcmc_y. Returns a tidy data frame of MCMC draws. Omitted if draws = FALSE.

x_summary_y: extract posterior summaries from x_mcmc_y. Returns a tidy data frame of MCMC draws. Omitted if summary = FALSE.

x_dic: extract deviance information criterion (DIC) info from x_mcmc_y. Returns a tidy data frame of DIC info. Omitted if dic = FALSE.

Details

The MCMC targets use R2jags::jags() if n.cluster is 1 and R2jags::jags.parallel() otherwise. Most arguments to tar_jags() are forwarded to these functions.

Target objects

Most stantargets functions are target factories, which means they return target objects or lists of target objects. Target objects represent skippable steps of the analysis pipeline as described at https://books.ropensci.org/targets/. Please read the walkthrough at https://books.ropensci.org/targets/walkthrough.html to understand the role of target objects in analysis pipelines.

For developers, https://wlandau.github.io/targetopia/contributing.html#target-factories explains target factories (functions like this one which generate targets) and the design specification at https://books.ropensci.org/targets-design/ details the structure and composition of target objects.

Examples

if (requireNamespace("R2jags", quietly = TRUE)) {
targets::tar_dir({ # tar_dir() runs code from a temporary directory.
targets::tar_script({
library(jagstargets)
# Do not use a temp file for a real project
# or else your targets will always rerun.
tmp <- tempfile(pattern = "", fileext = ".jags")
tar_jags_example_file(tmp)
list(
  tar_jags(
    your_model,
    jags_files = tmp,
    data = tar_jags_example_data(),
    parameters.to.save = "beta",
    stdout = R.utils::nullfile(),
    stderr = R.utils::nullfile()
  )
)
}, ask = FALSE)
targets::tar_make()
})
}
#> ▶ dispatched target your_model_file_X7184902e1d6
#> ● completed target your_model_file_X7184902e1d6 [0 seconds]
#> ▶ dispatched target your_model_data
#> ● completed target your_model_data [0 seconds]
#> ▶ dispatched target your_model_lines_X7184902e1d6
#> ● completed target your_model_lines_X7184902e1d6 [0.001 seconds]
#> ▶ dispatched target your_model_mcmc_X7184902e1d6
#> ● completed target your_model_mcmc_X7184902e1d6 [0.028 seconds]
#> ▶ dispatched target your_model_draws_X7184902e1d6
#> ● completed target your_model_draws_X7184902e1d6 [0.001 seconds]
#> ▶ dispatched target your_model_dic_X7184902e1d6
#> ● completed target your_model_dic_X7184902e1d6 [0.002 seconds]
#> ▶ dispatched target your_model_summary_X7184902e1d6
#> ● completed target your_model_summary_X7184902e1d6 [0.028 seconds]
#> ▶ ended pipeline [0.153 seconds]
#> Warning message:
#> package ‘targets’ was built under R version 4.3.3 
#>