Tidy posterior summaries from multiple MCMCs per model.

Run multiple MCMCs on simulated datasets and return posterior summaries and the effective number of parameters for each run.

tar_jags_rep_summary(
  name,
  jags_files,
  parameters.to.save,
  data = list(),
  data_copy = character(0),
  data_omit = character(0),
  variables = NULL,
  summaries = NULL,
  summary_args = NULL,
  batches = 1L,
  reps = 1L,
  combine = TRUE,
  n.cluster = 1,
  n.chains = 3,
  n.iter = 2000,
  n.burnin = as.integer(n.iter/2),
  n.thin = 1,
  jags.module = c("glm", "dic"),
  inits = NULL,
  RNGname = c("Wichmann-Hill", "Marsaglia-Multicarry", "Super-Duper",
    "Mersenne-Twister"),
  jags.seed = 123,
  log = NULL,
  progress.bar = "text",
  refresh = 0,
  tidy_eval = targets::tar_option_get("tidy_eval"),
  packages = targets::tar_option_get("packages"),
  library = targets::tar_option_get("library"),
  error = targets::tar_option_get("error"),
  memory = "transient",
  garbage_collection = targets::tar_option_get("garbage_collection"),
  deployment = targets::tar_option_get("deployment"),
  priority = targets::tar_option_get("priority"),
  resources = targets::tar_option_get("resources"),
  storage = targets::tar_option_get("storage"),
  retrieval = targets::tar_option_get("retrieval"),
  cue = targets::tar_option_get("cue")
)

Arguments

name
Symbol, base name for the collection of targets. Serves as a prefix for target names.

jags_files
Character vector of JAGS model files. If you supply multiple files, each model will run on the one shared dataset generated by the code in data. If you supply an unnamed vector, tools::file_path_sans_ext(basename(jags_files)) will be used as target name suffixes. If jags_files is a named vector, the suffixed will come from names(jags_files).

parameters.to.save
character vector of the names of the parameters to save which should be monitored.

data
Code to generate the data for the JAGS model.

data_copy
Character vector of names of scalars in data. These values will be inserted as columns in the output data frame for each rep. Useful for simulation studies where you want to check the results against some "true value" in the data. See the mcmc_rep vignette for an example.

data_omit
Elements of the data to omit from the JAGS model. Useful if you want to pass elements to data_copy without supplying them to the model. (Superfluous data causes warnings in JAGS.)

variables
Character vector of model parameter names. The output posterior summaries are restricted to these variables.

summaries
List of summary functions passed to ... in posterior::summarize_draws() through $summary() on the CmdStanFit object.

summary_args
List of summary function arguments passed to .args in posterior::summarize_draws() through $summary() on the CmdStanFit object.

batches
Number of batches. Each batch runs a model reps times.

reps
Number of replications per batch. Ideally, each rep should produce its own random dataset using the code supplied to data.

combine
Logical, whether to create a target to combine all the model results into a single data frame downstream. Convenient, but duplicates data.

n.cluster
number of clusters to use to run parallel chains. Default equals n.chains.

n.chains
number of Markov chains (default: 3)

n.iter
number of total iterations per chain (including burn in; default: 2000)

n.burnin
length of burn in, i.e. number of iterations to discard at the beginning. Default is n.iter/2, that is, discarding the first half of the simulations. If n.burnin is 0, jags() will run 100 iterations for adaption.

n.thin
thinning rate. Must be a positive integer. Set n.thin > 1 to save memory and computation time if n.iter is large. Default is max(1, floor(n.chains * (n.iter-n.burnin) / 1000)) which will only thin if there are at least 2000 simulations.

jags.module
the vector of jags modules to be loaded. Default are “glm” and “dic”. Input NULL if you don't want to load any jags module.

inits
a list with n.chains elements; each element of the list is itself a list of starting values for the BUGS model, or a function creating (possibly random) initial values. If inits is NULL, JAGS will generate initial values for parameters.

RNGname
the name for random number generator used in JAGS. There are four RNGS supplied by the base moduale in JAGS: Wichmann-Hill, Marsaglia-Multicarry, Super-Duper, Mersenne-Twister

jags.seed
random seed for JAGS, default is 123. This function is used for jags.parallell() and does not work for jags(). Use set.seed() instead if you want to produce identical result with jags()

log
Character of length 1, file path to write the stdout stream of the model when it runs. Set to NULL to print to the console. Set to R.utils::nullfile() to completely suppress all output. Does not apply to messages, warnings, or errors.

progress.bar
type of progress bar. Possible values are “text”, “gui”, and “none”. Type “text” is displayed on the R console. Type “gui” is a graphical progress bar in a new window. The progress bar is suppressed if progress.bar is “none”

refresh
refresh frequency for progress bar, default is n.iter/50

tidy_eval
Logical, whether to enable tidy evaluation when interpreting command and pattern. If TRUE, you can use the "bang-bang" operator !! to programmatically insert the values of global objects.

packages
Character vector of packages to load right before the target builds. Use tar_option_set() to set packages globally for all subsequent targets you define.

library
Character vector of library paths to try when loading packages.

error
Character of length 1, what to do if the target runs into an error. If "stop", the whole pipeline stops and throws an error. If "continue", the error is recorded, but the pipeline keeps going.

memory
Character of length 1, memory strategy. If "persistent", the target stays in memory until the end of the pipeline (unless storage is "worker", in which case targets unloads the value from memory right after storing it in order to avoid sending copious data over a network). If "transient", the target gets unloaded after every new target completes. Either way, the target gets automatically loaded into memory whenever another target needs the value. For cloud-based dynamic files such as format = "aws_file", this memory policy applies to temporary local copies of the file in _targets/scratch/": "persistent" means they remain until the end of the pipeline, and "transient" means they get deleted from the file system as soon as possible. The former conserves bandwidth, and the latter conserves local storage.

garbage_collection
Logical, whether to run base::gc() just before the target runs.

deployment
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "worker", the target builds on a parallel worker. If "main", the target builds on the host machine / process managing the pipeline.

priority
Numeric of length 1 between 0 and 1. Controls which targets get deployed first when multiple competing targets are ready simultaneously. Targets with priorities closer to 1 get built earlier. Only applies to tar_make_future() and tar_make_clustermq() (not tar_make()). tar_make_future() with no extra settings is a drop-in replacement for tar_make() in this case.

resources
A named list of computing resources. Uses:

Template file wildcards for future::future() in tar_make_future().

Template file wildcards clustermq::workers() in tar_make_clustermq().

Custom target-level future::plan(), e.g. resources = list(plan = future.callr::callr).

Custom curl handle if format = "url", e.g. resources = list(handle = curl::new_handle()).

Custom preset for qs::qsave() if format = "qs", e.g. resources = list(handle = "archive").

Custom compression level for fst::write_fst() if format is "fst", "fst_dt", or "fst_tbl", e.g. resources = list(compress = 100).

AWS bucket and prefix for the "aws_" formats, e.g. resources = list(bucket = "your-bucket", prefix = "folder/name"). bucket is required for AWS formats. See the cloud computing chapter of the manual for details.

storage
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "main", the target's return value is sent back to the host machine and saved locally. If "worker", the worker saves the value.

retrieval
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "main", the target's dependencies are loaded on the host machine and sent to the worker before the target builds. If "worker", the worker loads the targets dependencies.

cue
An optional object from tar_cue() to customize the rules that decide whether the target is up to date.

Value

tar_jags_rep_dic(name = x, jags_files = "y.jags") returns a list of targets::tar_target() objects:

x_file_y: reproducibly track the jags model file.

x_lines_y: contents of the jags model file.

x_data: dynamic branching target with simulated datasets.

x_y: dynamic branching target with tidy data frames of posterior summaries.

x: combine all the model-specific summary targets into a single data frame with columns to distinguish among the models. Suppressed if combine is FALSE.

Target objects represent skippable steps of the analysis pipeline as described at https://books.ropensci.org/targets/. Please see the design specification at https://books.ropensci.org/targets-design/ to learn about the structure and composition of target objects.

Details

The MCMC targets use R2jags::jags() if n.cluster is 1 and R2jags::jags.parallel() otherwise. Most arguments to tar_jags() are forwarded to these functions.

Examples

targets::tar_dir({
targets::tar_script({
library(jagstargets)
# Do not use a temp file for a real project
# or else your targets will always rerun.
tmp <- tempfile()
tar_jags_example_file(tmp)
list(
  tar_jags_rep_summary(
    your_model,
    jags_files = tmp,
    data = tar_jags_example_data(true_params = FALSE),
    parameters.to.save = "beta",
    batches = 2,
    reps = 2,
    log = R.utils::nullfile()
  )
)
}, ask = FALSE)
targets::tar_make()
})

#> ● run target your_model_file_file92f59947485
#> ● run target your_model_batch
#> ● run target your_model_lines_file92f59947485
#> ● run branch your_model_data_7b1b1358
#> ● run branch your_model_data_7b6d503e
#> ● run branch your_model_file92f59947485_24769822
#> ● run branch your_model_file92f59947485_4e151279
#> ● run target your_model
#> ● end pipeline
#>