One MCMC per model with multiple outputs

Targets to run a JAGS model once with MCMC and save multiple outputs.

tar_jags(
  name,
  jags_files,
  parameters.to.save,
  data = list(),
  summaries = list(),
  summary_args = list(),
  n.cluster = 1,
  n.chains = 3,
  n.iter = 2000,
  n.burnin = as.integer(n.iter/2),
  n.thin = 1,
  jags.module = c("glm", "dic"),
  inits = NULL,
  RNGname = c("Wichmann-Hill", "Marsaglia-Multicarry", "Super-Duper",
    "Mersenne-Twister"),
  jags.seed = 1,
  stdout = NULL,
  stderr = NULL,
  progress.bar = "text",
  refresh = 0,
  draws = TRUE,
  summary = TRUE,
  dic = TRUE,
  tidy_eval = targets::tar_option_get("tidy_eval"),
  packages = targets::tar_option_get("packages"),
  library = targets::tar_option_get("library"),
  error = targets::tar_option_get("error"),
  memory = targets::tar_option_get("memory"),
  garbage_collection = targets::tar_option_get("garbage_collection"),
  deployment = targets::tar_option_get("deployment"),
  priority = targets::tar_option_get("priority"),
  resources = targets::tar_option_get("resources"),
  storage = targets::tar_option_get("storage"),
  retrieval = targets::tar_option_get("retrieval"),
  cue = targets::tar_option_get("cue")
)

Arguments

name
Symbol, base name for the collection of targets. Serves as a prefix for target names.

jags_files
Character vector of JAGS model files. If you supply multiple files, each model will run on the one shared dataset generated by the code in data. If you supply an unnamed vector, tools::file_path_sans_ext(basename(jags_files)) will be used as target name suffixes. If jags_files is a named vector, the suffixed will come from names(jags_files).

parameters.to.save
character vector of the names of the parameters to save which should be monitored.

data
Code to generate the data list for the JAGS model. Optionally include a .join_data element to join parts of the data to correspondingly named parameters in the summary output. See the vignettes for details.

summaries
List of summary functions passed to ... in posterior::summarize_draws() through $summary() on the CmdStanFit object.

summary_args
List of summary function arguments passed to .args in posterior::summarize_draws() through $summary() on the CmdStanFit object.

n.cluster
number of clusters to use to run parallel chains. Default equals n.chains.

n.chains
number of Markov chains (default: 3)

n.iter
number of total iterations per chain (including burn in; default: 2000)

n.burnin
length of burn in, i.e. number of iterations to discard at the beginning. Default is n.iter/2, that is, discarding the first half of the simulations. If n.burnin is 0, jags() will run 100 iterations for adaption.

n.thin
thinning rate. Must be a positive integer. Set n.thin > 1 to save memory and computation time if n.iter is large. Default is max(1, floor(n.chains * (n.iter-n.burnin) / 1000)) which will only thin if there are at least 2000 simulations.

jags.module
the vector of jags modules to be loaded. Default are “glm” and “dic”. Input NULL if you don't want to load any jags module.

inits
a list with n.chains elements; each element of the list is itself a list of starting values for the BUGS model, or a function creating (possibly random) initial values. If inits is NULL, JAGS will generate initial values for parameters.

RNGname
the name for random number generator used in JAGS. There are four RNGS supplied by the base moduale in JAGS: Wichmann-Hill, Marsaglia-Multicarry, Super-Duper, Mersenne-Twister

jags.seed
random seed for JAGS, default is 123. This function is used for jags.parallell() and does not work for jags(). Use set.seed() instead if you want to produce identical result with jags()

stdout
Character of length 1, file path to write the stdout stream of the model when it runs. Set to NULL to print to the console. Set to R.utils::nullfile() to suppress stdout. Does not apply to messages, warnings, or errors.

stderr
Character of length 1, file path to write the stderr stream of the model when it runs. Set to NULL to print to the console. Set to R.utils::nullfile() to suppress stderr. Does not apply to messages, warnings, or errors.

progress.bar
type of progress bar. Possible values are “text”, “gui”, and “none”. Type “text” is displayed on the R console. Type “gui” is a graphical progress bar in a new window. The progress bar is suppressed if progress.bar is “none”

refresh
refresh frequency for progress bar, default is n.iter/50

draws
Logical, whether to create a target for posterior draws. Saves draws as a compressed posterior::as_draws_df() tibble. Convenient, but duplicates storage.

summary
Logical, whether to create a target to store a small data frame of posterior summary statistics and convergence diagnostics.

dic
Logical, whether to create a target with deviance information criterion (DIC) results.

tidy_eval
Logical, whether to enable tidy evaluation when interpreting command and pattern. If TRUE, you can use the "bang-bang" operator !! to programmatically insert the values of global objects.

packages
Character vector of packages to load right before the target builds. Use tar_option_set() to set packages globally for all subsequent targets you define.

library
Character vector of library paths to try when loading packages.

error
Character of length 1, what to do if the target runs into an error. If "stop", the whole pipeline stops and throws an error. If "continue", the error is recorded, but the pipeline keeps going. error = "workspace" is just like error = "stop" except targets saves a special workspace file to support interactive debugging outside the pipeline. (Visit https://books.ropensci.org/targets/debugging.html to learn how to debug targets using saved workspaces.)

memory
Character of length 1, memory strategy. If "persistent", the target stays in memory until the end of the pipeline (unless storage is "worker", in which case targets unloads the value from memory right after storing it in order to avoid sending copious data over a network). If "transient", the target gets unloaded after every new target completes. Either way, the target gets automatically loaded into memory whenever another target needs the value. For cloud-based dynamic files such as format = "aws_file", this memory policy applies to temporary local copies of the file in _targets/scratch/": "persistent" means they remain until the end of the pipeline, and "transient" means they get deleted from the file system as soon as possible. The former conserves bandwidth, and the latter conserves local storage.

garbage_collection
Logical, whether to run base::gc() just before the target runs.

deployment
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "worker", the target builds on a parallel worker. If "main", the target builds on the host machine / process managing the pipeline.

priority
Numeric of length 1 between 0 and 1. Controls which targets get deployed first when multiple competing targets are ready simultaneously. Targets with priorities closer to 1 get built earlier (and polled earlier in tar_make_future()). Only applies to tar_make_future() and tar_make_clustermq() (not tar_make()). tar_make_future() with no extra settings is a drop-in replacement for tar_make() in this case.

resources
A named list of computing resources. Uses:

Template file wildcards for future::future() in tar_make_future().

Template file wildcards clustermq::workers() in tar_make_clustermq().

Custom target-level future::plan(), e.g. resources = list(plan = future.callr::callr).

Custom curl handle if format = "url", e.g. resources = list(handle = curl::new_handle(nobody = TRUE)). In custom handles, most users should manually set nobody = TRUE so targets does not download the entire file when it only needs to check the time stamp and ETag.

Custom preset for qs::qsave() if format = "qs", e.g. resources = list(handle = "archive").

Arguments compression and compression_level to arrow::write_feather() and arrow:write_parquet() if format is "feather", "parquet", "aws_feather", or "aws_parquet".

Custom compression level for fst::write_fst() if format is "fst", "fst_dt", or "fst_tbl", e.g. resources = list(compress = 100).

AWS bucket and prefix for the "aws_" formats, e.g. resources = list(bucket = "your-bucket", prefix = "folder/name"). bucket is required for AWS formats. See the cloud computing chapter of the manual for details.

storage
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "main", the target's return value is sent back to the host machine and saved locally. If "worker", the worker saves the value.

retrieval
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "main", the target's dependencies are loaded on the host machine and sent to the worker before the target builds. If "worker", the worker loads the targets dependencies.

cue
An optional object from tar_cue() to customize the rules that decide whether the target is up to date.

Value

tar_jags() returns list of target objects. See the "Target objects" section for background. The target names use the name argument as a prefix, and the individual elements of jags_files appear in the suffixes where applicable. As an example, the specific target objects returned by tar_jags(name = x, jags_files = "y.jags", ...) returns a list of targets::tar_target() objects:

x_file_y: reproducibly track the JAGS model file. Returns a character vector of length 1 with the path to the JAGS model file.

x_lines_y: read the contents of the JAGS model file for safe transport to parallel workers. Returns a character vector of lines in the model file.

x_data: run the R expression in the data argument to produce a JAGS dataset for the model. Returns a JAGS data list.

x_mcmc_y: run MCMC on the model and dataset. Returns an rjags object from R2jags with all the MCMC results.

x_draws_y: extract posterior samples from x_mcmc_y. Returns a tidy data frame of MCMC draws. Omitted if draws = FALSE.

x_summary_y: extract posterior summaries from x_mcmc_y. Returns a tidy data frame of MCMC draws. Omitted if summary = FALSE.

x_dic: extract deviance information criterion (DIC) info from x_mcmc_y. Returns a tidy data frame of DIC info. Omitted if dic = FALSE.

Details

The MCMC targets use R2jags::jags() if n.cluster is 1 and R2jags::jags.parallel() otherwise. Most arguments to tar_jags() are forwarded to these functions.

Target objects

Most stantargets functions are target factories, which means they return target objects or lists of target objects. Target objects represent skippable steps of the analysis pipeline as described at https://books.ropensci.org/targets/. Please read the walkthrough at https://books.ropensci.org/targets/walkthrough.html to understand the role of target objects in analysis pipelines.

For developers, https://wlandau.github.io/targetopia/contributing.html#target-factories explains target factories (functions like this one which generate targets) and the design specification at https://books.ropensci.org/targets-design/ details the structure and composition of target objects.

Examples

targets::tar_dir({ # tar_dir() runs code from a temporary directory.
targets::tar_script({
library(jagstargets)
# Do not use a temp file for a real project
# or else your targets will always rerun.
tmp <- tempfile(pattern = "", fileext = ".jags")
tar_jags_example_file(tmp)
list(
  tar_jags(
    your_model,
    jags_files = tmp,
    data = tar_jags_example_data(),
    parameters.to.save = "beta",
    stdout = R.utils::nullfile(),
    stderr = R.utils::nullfile()
  )
)
}, ask = FALSE)
targets::tar_make()
})
#> • start target your_model_data
#> • built target your_model_data
#> • start target your_model_file_e927874e02e
#> • built target your_model_file_e927874e02e
#> • start target your_model_lines_e927874e02e
#> • built target your_model_lines_e927874e02e
#> • start target your_model_mcmc_e927874e02e
#> • built target your_model_mcmc_e927874e02e
#> • start target your_model_dic_e927874e02e
#> • built target your_model_dic_e927874e02e
#> • start target your_model_summary_e927874e02e
#> • built target your_model_summary_e927874e02e
#> • start target your_model_draws_e927874e02e
#> • built target your_model_draws_e927874e02e
#> • end pipeline
#>